+++ documentation/content/docs/(column)/demos/column-field-function_x/components.tsx
@@ -1,0 +1,312 @@
+import type { stockData } from "@1771technologies/grid-sample-data/stock-data-smaller";
+import { memo, type ReactNode } from "react";
+import { logos } from "@1771technologies/grid-sample-data/stock-data-smaller";
+import type {
+  CellRendererParams,
+  HeaderCellRendererParams,
+} from "@1771technologies/lytenyte-pro/types";
+import type { ClassValue } from "clsx";
+import clsx from "clsx";
+import { twMerge } from "tailwind-merge";
+
+type StockData = (typeof stockData)[number];
+
+export function tw(...c: ClassValue[]) {
+  return twMerge(clsx(...c));
+}
+
+function AnalystRatingCellImpl({ grid, row, column }: CellRendererParams<StockData>) {
+  const field = grid.api.columnField(column, row) as string;
+
+  let Icon: (() => ReactNode) | null = null;
+  const label = field || "-";
+  let clx = "";
+  if (label === "Strong buy") {
+    Icon = StrongBuy;
+    clx = "text-green-500";
+  } else if (label === "Strong Sell") {
+    Icon = StrongSell;
+    clx = "text-red-500";
+  } else if (label === "Neutral") {
+    Icon = Minus;
+  } else if (label === "Buy") {
+    Icon = Buy;
+    clx = "text-green-500";
+  } else if (label === "Sell") {
+    Icon = Sell;
+    clx = "text-red-500";
+  }
+
+  const aggModel = grid.state.aggModel.useValue();
+
+  const activeAgg = aggModel[column.id];
+  const isCount = activeAgg?.fn === "count";
+  const isGroupRow = grid.api.rowIsGroup(row);
+
+  return (
+    <div
+      className={tw(
+        "grid h-full grid-cols-[16px_1fr] items-center gap-4 text-nowrap px-3",
+        clx,
+        isGroupRow && isCount && "flex justify-end",
+      )}
+    >
+      {Icon && <Icon />}
+      <div>
+        {label}
+        {isCount && <span className="pl-1">Ratings</span>}
+      </div>
+    </div>
+  );
+}
+
+function Minus() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+      strokeWidth={1.5}
+      stroke="currentColor"
+      width={16}
+    >
+      <path strokeLinecap="round" strokeLinejoin="round" d="M5 12h14" />
+    </svg>
+  );
+}
+
+function Sell() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+      strokeWidth={1.5}
+      stroke="currentColor"
+      width={16}
+    >
+      <path strokeLinecap="round" strokeLinejoin="round" d="m19.5 8.25-7.5 7.5-7.5-7.5" />
+    </svg>
+  );
+}
+
+function Buy() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+      strokeWidth={1.5}
+      stroke="currentColor"
+      width={16}
+    >
+      <path strokeLinecap="round" strokeLinejoin="round" d="m4.5 15.75 7.5-7.5 7.5 7.5" />
+    </svg>
+  );
+}
+
+function StrongSell() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+      strokeWidth={1.5}
+      stroke="currentColor"
+      width={16}
+    >
+      <path
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        d="m4.5 5.25 7.5 7.5 7.5-7.5m-15 6 7.5 7.5 7.5-7.5"
+      />
+    </svg>
+  );
+}
+
+function StrongBuy() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+      strokeWidth={1.5}
+      stroke="currentColor"
+      width={16}
+    >
+      <path strokeLinecap="round" strokeLinejoin="round" d="m4.5 18.75 7.5-7.5 7.5 7.5" />
+      <path strokeLinecap="round" strokeLinejoin="round" d="m4.5 12.75 7.5-7.5 7.5 7.5" />
+    </svg>
+  );
+}
+
+export const AnalystRatingCell = memo(AnalystRatingCellImpl);
+
+function CompactNumberCellImpl({ row, grid, column }: CellRendererParams<StockData>) {
+  const field = grid.api.columnField(column, row);
+
+  const isGroup = grid.api.rowIsGroup(row);
+  const aggregations = grid.state.aggModel.useValue();
+  const isCount = aggregations[column.id]?.fn === "count";
+
+  const [label, suffix] =
+    typeof field === "number"
+      ? isGroup && isCount
+        ? [Math.round(field), ""]
+        : formatCompactNumber(field)
+      : ["-", ""];
+
+  return (
+    <div className="flex h-full w-full items-center justify-end gap-1 text-nowrap px-3 tabular-nums">
+      <span>{label}</span>
+      <span className="font-semibold">{suffix}</span>
+    </div>
+  );
+}
+
+export const CompactNumberCell = memo(CompactNumberCellImpl);
+function formatCompactNumber(n: number) {
+  const suffixes = ["", "K", "M", "B", "T"];
+  let magnitude = 0;
+  let num = Math.abs(n);
+
+  while (num >= 1000 && magnitude < suffixes.length - 1) {
+    num /= 1000;
+    magnitude++;
+  }
+
+  const decimals = 2;
+  const formatted = num.toFixed(decimals);
+
+  return [`${n < 0 ? "-" : ""}${formatted}`, suffixes[magnitude]];
+}
+
+const formatter = new Intl.NumberFormat("en-US", {
+  minimumFractionDigits: 2,
+  maximumFractionDigits: 2,
+});
+
+function CurrencyCellImpl({ grid, row, column }: CellRendererParams<StockData>) {
+  const field = grid.api.columnField(column, row);
+
+  const label = typeof field === "number" ? formatter.format(field) : "-";
+  const currency = "USD";
+
+  const aggregations = grid.state.aggModel.useValue();
+  const isCount = aggregations[column.id]?.fn === "count";
+
+  return (
+    <div className="flex h-full w-full items-center justify-end text-nowrap px-3 tabular-nums">
+      <div className="flex items-baseline gap-1">
+        <span>{label}</span>
+        {!isCount && <span className="text-[9px]">{currency}</span>}
+      </div>
+    </div>
+  );
+}
+
+export const CurrencyCell = memo(CurrencyCellImpl);
+
+function CurrencyCellGBPImpl({ grid, row, column }: CellRendererParams<StockData>) {
+  const field = grid.api.columnField(column, row);
+
+  const label = typeof field === "number" ? formatter.format(field) : "-";
+  const currency = "GBP";
+
+  const aggregations = grid.state.aggModel.useValue();
+  const isCount = aggregations[column.id]?.fn === "count";
+
+  return (
+    <div className="flex h-full w-full items-center justify-end text-nowrap px-3 tabular-nums">
+      <div className="flex items-baseline gap-1">
+        <span>{label}</span>
+        {!isCount && <span className="text-[9px]">{currency}</span>}
+      </div>
+    </div>
+  );
+}
+
+export const CurrencyCellGBP = memo(CurrencyCellGBPImpl);
+
+function PercentCellImpl({ grid, row, column }: CellRendererParams<StockData>) {
+  const field = grid.api.columnField(column, row) as number;
+
+  const isGroup = grid.api.rowIsGroup(row);
+  const aggregations = grid.state.aggModel.useValue();
+  const isCount = aggregations[column.id]?.fn === "count";
+
+  const label =
+    typeof field === "number"
+      ? isGroup && isCount
+        ? Math.round(field)
+        : formatter.format(field) + "%"
+      : "-";
+
+  return (
+    <div
+      className={tw("flex h-full w-full items-center justify-end text-nowrap px-3 tabular-nums")}
+    >
+      {label}
+    </div>
+  );
+}
+
+export const PercentCell = memo(PercentCellImpl);
+
+function SymbolCellImpl({ grid, row, column }: CellRendererParams<StockData>) {
+  if (grid.api.rowIsGroup(row)) {
+    const data = row.data[column.id];
+
+    const agg = grid.state.aggModel.get()[column.id];
+
+    return (
+      <div className="flex h-full items-center text-nowrap px-3">
+        {data as string} {agg.fn === "count" ? (data === 1 ? "Symbol" : "Symbols") : ""}
+      </div>
+    );
+  }
+  if (!grid.api.rowIsLeaf(row)) return null;
+
+  const symbol = grid.api.columnField(column, row) as string;
+  const desc = row.data?.[1];
+
+  return (
+    <div className="grid h-full w-full grid-cols-[32px_1fr] items-center gap-3 overflow-hidden text-nowrap px-3">
+      <div className="flex h-8 min-h-8 w-8 min-w-8 items-center justify-center overflow-hidden rounded-full">
+        <img
+          src={logos[symbol]}
+          alt=""
+          width={26}
+          height={26}
+          className="pointer-events-none h-[26px] min-h-[26px] w-[26px] min-w-[26] rounded-full bg-black p-1"
+        />
+      </div>
+      <div className="overflow-hidden text-ellipsis">{desc}</div>
+    </div>
+  );
+}
+export const SymbolCell = memo(SymbolCellImpl);
+
+export function HeaderRenderer({ column, grid }: HeaderCellRendererParams<StockData>) {
+  const name = column.name ?? column.id;
+
+  const isNumber = column.type === "number";
+
+  const aggModel = grid.state.aggModel.useValue();
+  const isGrouped = grid.state.rowGroupModel.useValue().length > 0;
+  const agg = (aggModel[column.id]?.fn ?? "") as string;
+
+  return (
+    <div
+      className={clsx(
+        "text-ln-gray-80 group relative flex h-full w-full items-center gap-1 overflow-hidden text-nowrap px-2 text-xs",
+        isNumber && "justify-end",
+      )}
+    >
+      <span className={tw(isNumber && "order-2")}>{name}</span>
+      {isGrouped && agg && (
+        <div className={tw("text-ln-primary-50 text-xs", isNumber && "order-1")}>({agg})</div>
+      )}
+    </div>
+  );
+}