+++ packages/lytenyte-core/src/header/header-group-cell.tsx
@@ -57,13 +57,13 @@ const HeaderGroupCellImpl = forwardRef<
       {...attrs}
       // Data attributes end
       style={{
-        ...props.style,
         ...styles,
         gridRow: "1 / 2",
         width,
         height: ctx.headerGroupHeight.useValue(),
         boxSizing: "border-box",
         opacity: (cell.isHiddenMove ?? false) ? 0 : undefined,
+        ...props.style,
       }}
     >
       {children == undefined ? cell.id : children}