+++ packages-auxiliary/eslint-1771/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @1771technologies/eslint-preset
 
+## 0.0.13
+
+### Patch Changes
+
+- verison
+
 ## 0.0.12
 
 ### Patch Changes