+++ packages-auxiliary/play-frame/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @1771technologies/play-frame
 
+## 0.0.18
+
+### Patch Changes
+
+- verison
+
 ## 0.0.17
 
 ### Patch Changes